requirements.txt
# Core dependencies
python-dotenv==1.0.1
pandas==2.2.2
requests==2.32.3

# Google AI
google-generativeai==0.3.2
langchain-google-genai  # No version constraint

# AutoGen
pyautogen==0.2.30

# LangChain + LangGraph
langchain==0.2.16
langchain-openai==0.2.8
langgraph==0.2.38

# CrewAI
crewai==0.28.8

# Optional
crewai-tools==0.1.6
